Sixth Commit
Finished the phone app with icons, etc. Also, changed the description
text/background to black because the phone's default seemed to be black.
I'm not sure how this will translate to actual user's phones, but the
screen shots were nice. Will need to change the values if I want to
update the desktop app.